+++ packages/core/src/domains/posts/index.ts
