+++ packages/core/src/domains/posts/queryPosts/types.ts
