+++ packages/core/src/domains/posts/queryTils/types.ts
