+++ packages/core/src/domains/posts/resources.ts
