+++ packages/core/src/domains/posts/utils.ts
