+++ packages/core/src/index.ts
@@ -1,3 +1,3 @@
 export * from './utils';
 export * from './types';
-export * from './domain';
+export * from './domains';
